cleanup docker setup

+++ compose/local/django/celery/beat/start.sh
@@ -1,10 +1,20 @@
 #!/usr/bin/env bash
 
+${ENV:?"Please set ENV non-empty"}
+
 set -o errexit
 set -o pipefail
 set -o nounset
-set -o xtrace
 
-
-rm -f './celerybeat.pid'
-celery -A seedorf.taskapp beat -l INFO
+if [ "$ENV" == "dev" ]
+then
+    set -o xtrace
+    rm -f './celerybeat.pid'
+    celery -A seedorf.taskapp beat -l INFO
+elif [ "$ENV" == "prd" ]
+then
+    celery -A seedorf.taskapp beat -l INFO
+else
+    echo "Invalid ENV variable"
+    exit
+fi
